Add files via upload
the example validation file for the NCI trees attached to this email. 
For the upgraded script, this will be generated for each of the trees and exported into a “Validation” folder, so one should be able to go through each of the files and cross check against the trees to make sure their outputs are correct.